+++ modules/core/m_sjoin.c
@@ -20,7 +20,7 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- *   $Id: m_sjoin.c,v 1.92 2001/05/26 00:48:17 davidt Exp $
+ *   $Id: m_sjoin.c,v 1.93 2001/05/28 13:45:36 jdc Exp $
  */
 #include "tools.h"
 #include "handlers.h"
@@ -254,8 +254,12 @@ static void ms_sjoin(struct Client *client_p,
 
   if (newts < 800000000)
     {
-      sendto_realops_flags(FLAGS_DEBUG, "*** Bogus TS %lu on %s ignored from %s",
-			   newts, chptr->chname, client_p->name);
+      sendto_realops_flags(FLAGS_DEBUG,
+			"*** Bogus TS %lu on %s ignored from %s",
+			(unsigned long) newts,
+			chptr->chname,
+			client_p->name);
+
       newts = oldts;
     }
 
@@ -369,8 +373,10 @@ static void ms_sjoin(struct Client *client_p,
       modebuf[1] = '\0';
     }
 
-  ircsprintf(buf, ":%s SJOIN %lu %s %s %s :", parv[0], tstosend, parv[2],
-          modebuf, parabuf);
+  ircsprintf(buf, ":%s SJOIN %lu %s %s %s :",
+	parv[0],
+	(unsigned long) tstosend,
+	parv[2], modebuf, parabuf);
 
   mbuf = modebuf;
   para[0] = para[1] = para[2] = para[3] = "";
