+++ modules/m_svinfo.c
@@ -20,7 +20,7 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- *   $Id: m_svinfo.c,v 1.21 2001/05/24 09:29:01 leeh Exp $
+ *   $Id: m_svinfo.c,v 1.22 2001/05/28 13:45:36 jdc Exp $
  */
 #include "handlers.h"
 #include "client.h"
@@ -109,13 +109,18 @@ static void ms_svinfo(struct Client *client_p, struct Client *source_p,
     {
       sendto_realops_flags(FLAGS_ALL,
        "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%d)",
-                 get_client_name(source_p, SHOW_IP),
-                 CurrentTime, theirtime, (int)deltat);
+		 get_client_name(source_p, SHOW_IP),
+		 (unsigned long) CurrentTime,
+		 (unsigned long) theirtime,
+		 (int) deltat);
 
       log(L_NOTICE,
        "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%d)",
                  get_client_name(source_p, SHOW_IP),
-                 CurrentTime, theirtime, (int)deltat);
+		 (unsigned long) CurrentTime,
+		 (unsigned long) theirtime,
+		 (int) deltat);
+
       exit_client(source_p, source_p, source_p, "Excessive TS delta");
       return;
     }
@@ -123,9 +128,11 @@ static void ms_svinfo(struct Client *client_p, struct Client *source_p,
   if (deltat > ConfigFileEntry.ts_warn_delta)
     { 
       sendto_realops_flags(FLAGS_ALL,
-                 "Link %s notable TS delta (my TS=%lu, their TS=%lu, delta=%d)",
-			   get_client_name(source_p, MASK_IP),
-			   CurrentTime, theirtime, (int)deltat);
+		"Link %s notable TS delta (my TS=%lu, their TS=%lu, delta=%d)",
+		get_client_name(source_p, MASK_IP),
+		(unsigned long) CurrentTime,
+		(unsigned long) theirtime,
+		(int) deltat);
     }
 }
